[OF-1728] feat!: user settings support avatar visibility information (#764)
* [OF-1728] feat: updated 'AvatarInfoResult' to support avatar selected state

* [OF-1728] feat!: update public API to include the avatar selected state
- included changes to ensure that the avatar selected state is serialized and deserialized

* [OF-1728] test: included new property in existing tests

* [OF-1728] refac: AvatarSelected - AvatarVisibility

This change is in service of client applications that want to express
whether their users' avatars should be visible or not within a space.

As a result, refactoring this change to be about avatar visibility
instead of whether an avatar has been selected.

* [OF-1728] refac: Avatar info - results parsing

Taking the opportunity to refactor the code responsible for
parsing the json string returned when avatar info settings
are requested.

This change has a few goals:
- Reduce early outs/divergent flow to improve readability.
- Improve commenting and code locality for readability.
- Improve error/log messaging.

*Note* - No actual changes to json schema, error handling or callback
patterns from a user perspective.

* [OF-1728] fix: typo

Corrected an overlooked lingering avatar-setting-related string.

* [OF-1728] fix: Avatars are by default visible

If a user has no avatar settings, we default to assuming
that their avatar should be visible (previously we defaulted
that they should be invisible).

* [OF-1728] test: avatar settings tests

Refactoring the current avatar info setting tests to
use the matrix-based approach, allowing us to express multiple tests
using various permutations of avatar settings succinctly.

Additionally, created a new test to validate that default avatar
settings are as expected when no settings data exists.

* [OF-1728] fix: removed const in Get/SetAvatarVisible

---------

Co-authored-by: MAG-ChristopherAtkinson <Christopher.Atkinson@magnopus.global>
Co-authored-by: Sam Birley <sam.birley@magnopus.com>