[OF-1689] refac!: Branch on patch logic in SpaceEntity
Don't just assume a patch based workflow, introduce a branch so SpaceEntity can behave "normally" in an offline state.

This is done by moving all the patch logic into another type, `SpaceEntityStatePatcher`, which serves as a holding ground for dirty property updates, as well as essentially being a representation of "A Patch"

This establishes a repeatable pattern for updating SpaceEntity properties. However, this is broken by two things at the moment:
- ParentID
- Updating properties on components

This is all accidental complexity imo, and I will likely address it in later commits.