[OF-1689] style: Remove List Aliases
It was becoming too burdensome to get these right when trying to keep public/private headers, as well as module boundarys in my head.

We can add these back in if folk insist, although we should do it as a proper `TypeAliases.h` file, rather than just putting these things in header files. Does not work well when your base type is in `Common` but your derived are in `Multiplayer`, and you want to avoid `std` dependencies in the base.