[OF-1689] refac: Remove patch concept from OfflineRealtimeEngine
Remove the concept of patches or pending operations from OfflineRealtimeEngine, making it fully syncronous, (to go with the SpaceEntity changes)

The methods are still implemented, I am yet to decide whether I want methods such as `ProcessPending` to be interface methods or not. Probably not, but it's worth thinking about client ergonomics some before deciding for real.

This involved some work to the utils. Methods MUST be named for what they do, It's very difficult to be able to work in a system otherwise.

There was a particularly nastily named function here, which I have renamed to `RemoveAsChildFromParent` to actually reflect what it does, as it was named the precise opposite.