[OF-1689] refac: Normalize parentID sequencing
We can't go the whole way and make parentID a regular dirtyable property, because of the wrapper gen (and probably also the CHS package format ... wish there was a spec)

This at least expresses ParentID such that the update sequencing is the same as the other properties, deferring to the patcher if neccesary, then coming back to the `Direct` set method as the final point of mutation.

Creates a bit of a weird nested optional to store that state, but workable.