[OF-1689] feat: Add GetAllEntities to RealtimeEngine interface
We needed this for the OfflineRealtimeEngine script integration, and it seems like an incredibly useful method. Not sure why it isn't a thing already.

Nonetheless, we're scaling, a warning is good enough for now. No need to solve problems we don't have.

Very annoying how the wrapper gen forces pointer returns here.