[OF-1689] test: Add basic test for offline conversations
Hmm ... an annoyance that conversations are special in that they interact with the signal r connection, almost completely missed that this was a case.

Added a tiny extension to the error message to make it a tad more obvious to someone who hasn't started the multiplayer connection, but we may want to consider our policy more strongly on this.

I think we can have read only conversations in a space, but it's sort of a big thing to consider.