[OF-1689] refac: Make the realtime engine responsible for whether it creates Patched entities.
This is quite conceptual. We want to make sure that the thing that answers the "Do space entities have patchers" question is the realtime engine.

We _also_ want to make sure SpaceEntities cannot be constructed in unexpected configs (like offline with patcher ... at least for the moment). This solves both problems.

Potentially somewhat temporary, but I wouldn't hate if it wasn't.