[OF-1750] feat: LoginAsGuestWithDeferredProfileCreation
Provide the special case API for logging into MCS in an "optimized" manner.

This is a unique endpoint for specific projects, and should not be called in the general case. A large warning is added to the documentation to that effect.

The endpoint has an undefined stretch of time before other profile related methods will become capable of returning non 404 responses. This is a shame, but it's been documented.

Need to figure this out, I understand that in distributed systems, a 404 is always a possibility, but there's a difference between an error always being possible vs a success never being possible. These are two categories imo and I would like to be able to know when endpoints are capable of returning success results vs when they are not. Otherwise you've just gotta sleep, loop, and hope.