[OF-1750] feat: LoginAsGuestWithDeferredProfileCreation (#789)
Provide the special case API for logging into MCS in an "optimized" manner.

This is a unique endpoint for specific projects, and should not be called in the general case. A large warning is added to the documentation to that effect. Does not provide an option to start a multiplayer connection, as it should not be used in that context.

The endpoint has an undefined stretch of time before other profile related methods will become capable of returning non 404 responses. This is a shame, but it's been documented.