[NT-0] fix: Resolve initialization ordering issue (#795)
* [NT-0] fix: Resolve initialization ordering issue

We had a circular dependency issue in the SystemsManager due to SignalR requiring the AuthContext from the UserSystem for construction. To get around this we pass nullptr to the UserSystem ctor for the NetworkEventBus, and then call a newly introduced  method to set it after the  NetworkEventBus has been constructed.