[OF-1739] test: Add batched analytics events (#791)
Added two new methods to the AnalyticsSystem; BatchAnalyticsEvent and FlushAnalyticsEvents.

The BatchAnalyticsEvent method makes it possible to add an analytic event to a queue to be sent to MCS later as part of a batch.

The queue will be processed when one of the following conditions is met:
1. The time since the last batch was sent reaches the batch rate (default 60 seconds).
2. The number of events in the queue reaches the maximum batch size (default 25 events).
3. The client application calls FlushAnalyticsEvents() as part of their log out or shut down procedure to force the batch to be sent.

The FlushAnalyticsEvents method will trigger immediate dispatch of the Analytics Records queue to MCS. This method is designed to be called as part of a client applications log out and shutdown procedures to ensure that any queued analytics records are flushed and sent to MCS before the user is logged out or the application is shut down.