[NT-0] fix: Duplication of asset collections
Following an update to the endpoints for duplicating prototypes
we updated the corresponding CSP abstraction.

However, as part of that we forgot to allocate memory when declaring
shared pointers that we then go on to derefrence.

Fixed by allocating memory and assigning it to the shared pointers when
invoking their constructors.