[OF-1733] refac!: Move settings objects to common
Need to do this because we need to put them on the login response, which is common.

It's a bit awkward, as we also have the result types which are `systems`. I'm thinking that result types should begin to be centralised into their own namespace.

As `ApplicationSettings` only had the result type stuff left in the file, I deleted it and put it into `ApplicationSettingsSystem` ... although honestly i'm not sure about this. It's non-symmetric with the already non-symettrically named `SettingsCollection`