[NT-0] test: Changed log category test pattern
Declaring log categories inline when invoking
checks like `EXPECT_CALL` (instead of using a locally declared var).

Makes it more obvious to the reader that this is the log
level we expect to be returned by the call.

Also fixed a variety of places where the wrong log
level  was being expected in tests. There may be more, we'll see.