[OF-1769] test: simplifying GetAgoraUserTokenTest
The test was validating log messages didn't contain specific
strings relating to errors, but we are already validating that the
response came back ok.

Also the way the the test was written was designed to force the test
to wait until callbacks timed out. Which is pointless wasted time.

Removed the log validation.