[OF-1769] refac!: Migrate external service proxy (#815)
We've had feedback that the external service proxy endpoints/functionality supported by CSP was not well known or understood.

With new external services being planned for integration with MCS, it's prescient to solve that problem and make it easier to find and use these endpoints.

We determined that the problem was a question of visibility. The functionality already existed, but it was buried in the user system (evolved from an initial specialized implementation in service of Agora) and the terms/names could have been clearer.

Refactoring our support for the external service proxy endpoint to make is simple and clear to use, via a new (small) CSP system - the ExternalServiceProxySystem.