[OF-1796] feat: Generic ExternalService result (#817)
It was pointed that we cannot assume any schema in the Json data
returned via the external service proxy services.

This is because the proxy service is assumed to simply acts as a router
for invoking operations on external services, which feeds the result
directly back to the caller.

To account for that, this change refactors the result class used by the
external service system into two classes. One generic class for
returning the result with no tampering, and another for the Agora
specialization.

The generic version is used by
ExternalServiceProxySystem::InvokeOperation, and returns the full string
value of the operationResult field in the returned payload.

The specialized class is used by
ExternalServiceProxySystem::GetAgoraUserToken. Since we can assume the
schema in this specialized instance, this class inherits from the
generic result class, takes the raw operationResult value, and parses
it for the token (this is what we had before for all calls).

The mock test has been updated to not assume any particular Json format.
Instead, it now simply stress tests whether the Json under the
operationResult is valid.

Some other small readability improvements have come along for the ride.