[OB-4737] fix: Script execution by elected leader
With the realtime engine refactor, we introduced a bug
where scripts would not be executed by the script leader
unless they were also the owner of the entity hosting the script.

This would manifest to users as scripts not working, until after they
touched any state on the entity (at which point CSP claims ownership).

However scripts would then appear to stop work again later, when
another user enters the space, becomes the script leader but once
again does not have ownership.

The fix is to restore the logic gating script execution; rather than
checking for entity ownership, we check to see if we are the elected
leader.

Non-online engines, however, should still continue to always run
scripts.