[OB-4737] fix: Scripts and disabled election
The API allows clients to express online realtime engines
with leadership election disabled. Which means we cannot
assume that when online is used, election will always be used.

Addressed by assuming by default that scripts may be run locally,
but if online and there is an election manager, then they can only
be run if the local client is the elected leader.