[OB-4420] fix: Add catch-all sentry workaround to allow Unity iOS to load old incompatible spaces (#819)
This is not at all a real fix, it's just a patch we can choose to apply that will allow Unity ios to load old spaces in the same way the other clients currently can. Putting this in based on manual testing on the device, as the worst that can happen is that it won't work. (read: low risk of behaviour change)

Testing this would imply either threading the log system down, inventing a whole new callback pattern, or even just moving this catch, which isn't something I'm willing to do for an off book fix like this.

We should discuss this in standup, not sure if we should merge this or not, just putting it up to have it ready.