[OB-4711] fix: Result parser for copied assets
Recently, the signature for the services function handling duplication
of asset collections was changed.

When we updated CSP to account for that, we missed that the DTOs
returned when invoking duplications had also changed...

To fix that, we now have a new child class of the
AssetCollectionsResult class (which is typically used to process and
represent arrays of asset collections returned to clients) and we
specialize the OnResponse function to reason about these new DTOs.

Doing it in this way means no breaking changes for client applications.