[OB-4711] fix: Asset duplications endpoint usage
We had a typo which meant that we were invoking asset
duplication without actually specifying the IDs of the assets
we were being asked to duplicate.

Came from a change to the signature of the services endpoint.
We'd already updated to account for that change, but crucially
missed that specifying IDs of assets to be duplicated was no longer
being done.

This change also fixes the expected Dto to be returned, and adopts
the new AssetCollectionsCopyResult class for processing the results.