[OB-4711] fix: Asset duplication (#832)
* [OB-4711] fix: Result parser for copied assets

Recently, the signature for the services function handling duplication
of asset collections was changed.

When we updated CSP to account for that, we missed that the DTOs
returned when invoking duplications had also changed...

To fix that, we now have a new child class of the
AssetCollectionsResult class (which is typically used to process and
represent arrays of asset collections returned to clients) and we
specialize the OnResponse function to reason about these new DTOs.

Doing it in this way means no breaking changes for client applications.

* [OB-4711] fix: Asset duplications endpoint usage

We had a typo which meant that we were invoking asset
duplication without actually specifying the IDs of the assets
we were being asked to duplicate.

Came from a change to the signature of the services endpoint.
We'd already updated to account for that change, but crucially
missed that specifying IDs of assets to be duplicated was no longer
being done.

This change also fixes the expected Dto to be returned, and adopts
the new AssetCollectionsCopyResult class for processing the results.

* [OB-4711] test: Reenable asset copy test

And corrected an invalid expectation about the number of
assets that will be copied. No idea why this would ever have been
expected to be two..

* [OB-4711] refac: Removed virtual qualifiers

As they were no longer needed.