[NT-0] fix: Uncaught exception in DTO parsing
CSP assumed certain fields would always be present on
prototype DTOs. Not unreasonable, but also a source of
uncaught exceptions if they weren't.

Fixed by following the same pattern as with other fields, and
only setting the properties following a check to see if the DTO has
them.

Also, removed the 'Charity' DTO type, as this was agreed in a recent
CSP working group session.