[OF-1784] feat: Adding Multiplayer Leadership election class
This object interfaces with the signalr hub methods to add more funcitonality for leader election. This object has been added to the OnlineRealtimeEngine to encapsulate this functionality, instead of just adding it to the OnlienRealtimeEngine directly. This commit does not add the necessary integration to get it working with scripts