[OF-1736] refac: Pass NetworkEventBus by ref
The SpaceSystem was being passed to the NetworkEventBus ctor by ptr and then we had a nullptr check to ensure it was valid before attempting to add an event listener. However, the NetworkEventBus cannot be null and should therefore be passed by ref.
The MultiplayerConnection::GetEventBus() method has been updated to return a reference to the EventBus. This ref is then passed to the systems, which dereference it before passing it to the SystemBase class. This is similar to what is done with the LogSystem.

I also took the opportunity to remove any 'In' prefixes from related method arguments.
[OF-1736] fix: Update