[NT-0] fix: Removing an asset collection assert (#841)
A CSP client app developer recently reported a crash when attempting to enter a space. Turns out it was due to there being an asset collection/prototype in the space that did not conform to the types CSP expects, which caused CSP to fire an assert.

To fix, we follow the same pattern as with Asset.cpp and emit an Error log message instead of an assert.