[OF-1804] feat: Add a new TeamCity script
As part of the nightly publish process we will be copying the .github/DEADME.md file to the package directory.

The plan is to replace our existing 'GenerateReadMeWithLink' script which runs as part of the Publish with this one. Once replaced the old script will be removed.

This script has been tested locally and will be validated in a temporary 'Test Publish CSP' TeamCity config as part of this initiaitve.