[OF-1804] ci: Add ssh deploy key
It was proving problematic to get the GitHub Action to successfully bypass the branch protection rules so that it could commit the changelog update to main without requiring a PR and approvals.

This issues has now been resolved by validating the required GitHub settings using my credentials and local testing. I then updated the GitHub settings to include the GitHub user and confirmed that the Action then performed correctly.

This local testing required that I then perform a reset to remove my test commit from main.
I also removed my last commit which was updating the GitHub Action to use a deploy token rather than the GitHub PAT. Running the Action again it did not work. I am therefore reinstating the Deploy Key as it is clearly required.