[OF-1784] feat: Integrate server-side leader election
This is a large peave of work. It contains - Integerating new server-side leader eleciton manager. You will notice a lot of conditions to check if server-side election is enaled. This is temporasry until we remove the client side election at a later date. It also contains the rafactored EnterSpace flow to check for server-side eleciton from the spaces scope, which is curently set by chs. It also registeres this scope with the server side eleciton manager.