[NT-0] fix: List::Insert erasing 0th element (#857)
Seems like it's been this way since the beginning of time.

The placement new (sort of redundant even now...) was erasing the 0th element no matter what.
Internally, we only ever use `Insert` at the 0th element, so it's not noticed.

Externally, this could very well be the cause of random seeming crashes. Depends on how much use this type has.

There's two other bugs in this function, todo with memcpy being undefined behaviour on non-trivial types, and the placement new assignment just being a memory leak for any types that need to run destructors.

Found this during the SWIG stuff, there were no unit tests for this type in CSP proper.