[OF-1784] Multiplayer System (#842)
* [OF-1784] fix: Multiplayer uri is now handled correctly

We were previously conflating the multiplayer hub uri and the multiplayer service uri, which require different values, but also a both have the same change to the RootUri. 2 seperate uris have now been setup to represent the 2

The helper function has also been updated which transforms the RootUri to be generic, so it can be used for both multiplayer uris

* [OF-1784] feat: Adding MultiplayerSystem

This adds a multiplayey system where we can put functionality from the chs multiplayer service api. We may want to make other system abstractions in the future, as this api contains a lot of different functionality. For now, only the functions needed for testing leader election have been implemented

* [OF-1784] docs: Add additional docs

* [OF-1784] test: Add test for getting scopes out of space

* [OF-1784] refac: updating apis to use unique_ptrs

* [OF-1784] refac: Update to use case with logging

* [OF-1784] docs: Adding comments to reference type

* [OF-1784] fix: Fixing initialization order

* [OF-1784] fix: Updated test function name

* [OF-1784] fix: Updated GetScopesBySpace to error if user is not in the space

* [OF-1784] docs: updated GroupId comment

* [OF-1784] test: Fixing test error code