[NT-0] chore: NO_EXPORT all OnResponse methods (#864)
* [NT-0] chore: NO_EXPORT all OnResponse methods

The ResultBase `OnResponse` is NO_EXPORT, so the overrides should be do. This is for SWIG, as we strip all the NO_EXPORT types away, but we still need it to compile, thus overriden methods need to be in the same annotation category as their base.

Some of the OnResponse on existing response handlers are already NO_EXPORT, but not all.

* [NT-0] fix: Update check-for-update to support forks

ref only provides enough information to fetch a commit local to the current repo, sha will allow forks.