Delete pre-commit hook. (#873)
In my opinion, this only hurts us. Yes it forces code to be formatted before it goes into the repo ... although clearly not really as _plenty_ of unformatted code gets in.

I don't doubt that removing this will increase the amount of weird whitespacing that gets into the codebase, and increase the level of whitespace churn in subsequent PRs, but this has caused _real work_ to get lost several times now, due to the automatic git-add.