[OF-1784] Server side election (#869)
* [OF-1784] feat: Added server-side leader election manager

This class is responsible for keeping track of leaders for each scope and sending heartbeats if the local client is the leader

* [OF-1784] feat: Added new election methods to hub map

* [OF-1784] feat: Refactored MultiplayerConnection to add continuations

This helps in a later step when we need to add additional calls to the EnterSpace flow

* [OF-1784] feat: Integrate server-side leader election

This is a large piece of work. It contains - Integerating new server-side leader eleciton manager. You will notice a lot of conditions to check if server-side election is enaled. This is temporasry until we remove the client side election at a later date. It also contains the rafactored EnterSpace flow to check for server-side eleciton from the spaces scope, which is curently set by chs. It also registeres this scope with the server side eleciton manager.

* [OF-1784] test: Added leader election integration test

* [OF-1784] test: Added script test

* [OF-1784] fix: fixing compile errors

* [OF-1784] feat: fixing tests

Update client-side leader election to be passed callback once initialized. This is due to the initialization location being changed. Also updated expected logs in tests due to EnterSpace refactor

* [OF-1784] fix: Updating ServersideLeaderElectionManager class

* [OF-1784] fix: Test namespace naming

* [OF-1784] fix: Spelling

* [OF-1784] docs: add comments to nullptr check

* [OF-1784] fix: Added error log

* [OF-1784] fix: Updating dto conversion to get client id

* [OF-1784] fix: Updated exception string

* [OF-1784] docs: updated comment

* [OF-1784] docs: updated comment

* [OF-1784] docs: add changelog

* [OF-1784] test: Updating log string

* [OF-1784] fix: nulling realtimeengine when exiting space

* [OF-1784] docs: fix spelling

* [OF-1784] fix: fix errors

* [OF-1784] fix: spelling

* [OF-1784] fix: spelling

* [OF-1784] docs: added comment fo heartbeat interval

* [OF-1784] fix: updating test to wait for callback