[NT-0] fix: Fix rare race condition in legacy leader election
When interacting with enable/disable leader election immediately after joining a space, there is a possibility for a race condition if the deletion happens after the check for null to branch the leader election, but before the actual access.

This occurs most readily in `AddSecondScriptTest`, where we explicitly disable leader election.