[NT-0] chore: Establish hashing pattern
Add std::hash specializations for our vector types, allowing them be stored in containers. The real reason for this however is so I can satisfy the IEquatable interface for C#.

This starts the pattern for these, we're going to need to do this for a fair chunk of our types, any that are conceptually "value-types".

We'll need one of these files for each of our 3 modules.
The reason this is centralized is twofold:

- std::hash specializations have to live in the std:: namespace so you don't get any information hiding by putting them along with the class anyhow
- I need to ignore these from the current wrapper gen, but _not_ for the new one. The only way I can really do this at the moment is on a file-by-file basis.